Introduce PolymorphismStyle.None
This implements #378 by introducing a new `None` style that omits the tag/property entirely from the output yaml.

Note this is asymmetric and reads are treated the same as `Tag` as reads require a tag.

I added a few tests to JvmYamlWritingTest for both this case and some added coverage for the existing two cases as well.